+++ server/mocks/statisticsMock.js
@@ -1,0 +1,16 @@
+const statisticsMock = [
+  {
+    CreationDate: "2021-06-04",
+    CountryCodeId: "MEX",
+    Deaths: "235000",
+    Confirmed: "603000",
+  },
+  {
+    CreationDate: "2021-06-04",
+    CountryCodeId: "COL",
+    Deaths: "235000",
+    Confirmed: "603000",
+  },
+];
+
+module.exports = statisticsMock;