+++ src/redux/actions/predictionsActions.js
@@ -1,0 +1,9 @@
+// Import const
+import { 
+  START_GET_DEATH_PREDICTIONS
+} from '../consts';
+
+export const getDeathPredictions = payload => ({
+  type: START_GET_DEATH_PREDICTIONS,
+  payload
+});